requirements.txt
pyqt5
pillow
openai
pyyaml
keyboard
pyperclip
